To create a chat box in Java, you can follow these steps
Create a GUI: You can use Java's Swing library to create a graphical user interface for your chat box. This would include creating a window, text input fields, and buttons for sending messages.

Implement the Networking Layer: You will need to implement the networking layer for communication between the client and the server. You can use Java's socket programming to create a connection between the client and the server and send messages back and forth.

Implement the Chat Functionality: You will need to implement the chat functionality by writing the logic to send and receive messages between the client and server. You will also need to store the messages in some data structure and display them in the chat box